@@ -20,7 +20,7 @@ Gem::Specification.new do |spec|
   spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
   spec.require_paths = ["lib"]
 
-  spec.add_dependency "rails",       ">= 4.2"
+  spec.add_dependency "rails",       "~> 5.0"
   spec.add_dependency "activeadmin", ">= 1.0"
   spec.add_dependency "devise",      "~> 4.3"
 
